+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.0.2 (2013/10/08)
 
 - Add `vagrant sakura-list-id` command.
+- Support `zone_id` configuration.
 
 # 0.0.1 (2013/10/07)